refactor: move templates and members into guild
