+++ packages/core/src/api/channel.ts
@@ -1,6 +1,11 @@
 /* eslint-disable jsdoc/check-param-names */
-import { makeURLSearchParams, type REST } from '@discordjs/rest';
+import { makeURLSearchParams, type REST, type RawFile } from '@discordjs/rest';
 import {
+	type RESTGetAPIChannelMessageReactionUsersQuery,
+	type RESTGetAPIChannelMessageReactionUsersResult,
+	type RESTPatchAPIChannelMessageResult,
+	type RESTPostAPIChannelMessageJSONBody,
+	type RESTPostAPIChannelMessageResult,
 	Routes,
 	type RESTDeleteAPIChannelResult,
 	type RESTGetAPIChannelInvitesResult,
@@ -22,6 +27,114 @@ import {
 
 export class ChannelsAPI {
 	public constructor(private readonly rest: REST) {}
+
+	/**
+	 * Sends a message in a channel
+	 *
+	 * @param channelId - The id of the channel to send the message in
+	 * @param message - The options to use when sending the message
+	 */
+	public async sendMessage(
+		channelId: Snowflake,
+		{ files, ...body }: RESTPostAPIChannelMessageJSONBody & { files?: RawFile[] },
+	) {
+		return this.rest.post(Routes.channelMessages(channelId), {
+			files,
+			body,
+		}) as Promise<RESTPostAPIChannelMessageResult>;
+	}
+
+	/**
+	 * Edits a message
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to edit
+	 * @param options - The options to use when editing the message
+	 */
+	public async editMessage(
+		channelId: Snowflake,
+		messageId: Snowflake,
+		{ files, ...body }: RESTPostAPIChannelMessageJSONBody & { files?: RawFile[] },
+	) {
+		return this.rest.patch(Routes.channelMessage(channelId, messageId), {
+			files,
+			body,
+		}) as Promise<RESTPatchAPIChannelMessageResult>;
+	}
+
+	/**
+	 * Fetches the reactions for a message
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to get the reactions for
+	 * @param emoji - The emoji to get the reactions for
+	 * @param options - The options to use when fetching the reactions
+	 */
+	public async getMessageReactions(
+		channelId: Snowflake,
+		messageId: Snowflake,
+		emoji: string,
+		options: RESTGetAPIChannelMessageReactionUsersQuery = {},
+	) {
+		return this.rest.get(Routes.channelMessageReaction(channelId, messageId, emoji), {
+			query: makeURLSearchParams(options as Record<string, unknown>),
+		}) as Promise<RESTGetAPIChannelMessageReactionUsersResult>;
+	}
+
+	/**
+	 * Deletes a reaction for the current user
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to delete the reaction for
+	 * @param emoji - The emoji to delete the reaction for
+	 */
+	public async deleteOwnMessageReaction(channelId: Snowflake, messageId: Snowflake, emoji: string) {
+		await this.rest.delete(Routes.channelMessageOwnReaction(channelId, messageId, emoji));
+	}
+
+	/**
+	 * Deletes a reaction for a user
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to delete the reaction for
+	 * @param emoji - The emoji to delete the reaction for
+	 * @param userID - The id of the user to delete the reaction for
+	 */
+	public async deleteUserMessageReaction(channelId: Snowflake, messageId: Snowflake, emoji: string, userID: Snowflake) {
+		await this.rest.delete(Routes.channelMessageUserReaction(channelId, messageId, emoji, userID));
+	}
+
+	/**
+	 * Deletes all reactions for a message
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to delete the reactions for
+	 */
+	public async deleteAllMessageReactions(channelId: Snowflake, messageId: Snowflake) {
+		await this.rest.delete(Routes.channelMessageAllReactions(channelId, messageId));
+	}
+
+	/**
+	 * Deletes all reactions of an emoji for a message
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to delete the reactions for
+	 * @param emoji - The emoji to delete the reactions for
+	 */
+	public async deleteAllMessageReactionsForEmoji(channelId: Snowflake, messageId: Snowflake, emoji: string) {
+		await this.rest.delete(Routes.channelMessageReaction(channelId, messageId, emoji));
+	}
+
+	/**
+	 * Adds a reaction to a message
+	 *
+	 * @param channelId - The id of the channel the message is in
+	 * @param messageId - The id of the message to add the reaction to
+	 * @param emoji - The emoji to add the reaction with
+	 */
+	public async addMessageReaction(channelId: Snowflake, messageId: Snowflake, emoji: string) {
+		await this.rest.put(Routes.channelMessageOwnReaction(channelId, messageId, emoji));
+	}
 
 	/**
 	 * Fetches a channel
